drivers: led_strip: add SK9822 LED driver.
The SK9822 LED has started to replace the harder to source
APA102C in most designs. The protocal is similar to the APA102C
but differs by having theend frame use zeros instead of ones.
This driver has been tested on a vmu_rt1170 (SPI EXT1) and
mr_canhubk344 (SPI2) with the led_strip sample.

Signed-off-by: Benjamin Perseghetti <bperseghetti@rudislabs.com>