bmm350: Add I3C bus support
Add bus support through RTIO bus transfers.

Disclaimer:

The existing driver has a known issue of being able to pull
the OTP compensation parameters, since the sensor makes them
unnaccessible after a power-on or soft-reset: therefore a soft-reset is
required to guarantee these will become readable on every reboot.

A possible solution is to send a soft-reset and then perform an I3C
re-attachment and continue the initialization process. However, neither
NXP nor ST (the hardware I possess) have this support on their
respective drivers, therefore it isn't implemented.

The implication of this issue not being solved is that software reboots
(after a power-on reset), will not contain the OTP values, therefore
data-readings from that point will not be properly compensated.

Signed-off-by: Luis Ubieda <luisf@croxel.com>