[TO BE UPSTREAMED] rtio: Add timeout option for consuming cqe items
As a way to provide error-handling to streaming clients (e.g: if no
data comes through in 100-ms assume it won't come vs block
indefinitely).

This could happen for various reasons (e.g: no CQE items left to give
back, or a bug in the RTIO IODEV code that causes it to omit the
result). In any case, we want to provide a forgiving way to recover
from these scenarios.

Signed-off-by: Luis Ubieda <luisf@croxel.com>